+++ pkg/tfgen/generate.go
@@ -248,6 +248,7 @@ type resourceType struct {
 	schema     *schema.Resource
 	info       *tfbridge.ResourceInfo
 	docURL     string
+	parsedDocs parsedDoc // parsed docs.
 }
 
 func (rt *resourceType) Name() string { return rt.name }
@@ -256,32 +257,34 @@ func (rt *resourceType) Doc() string  { return rt.doc }
 // IsProvider is true if this resource is a ProviderResource.
 func (rt *resourceType) IsProvider() bool { return rt.isProvider }
 
-func newResourceType(name, doc, docURL string, schema *schema.Resource, info *tfbridge.ResourceInfo,
+func newResourceType(name string, parsedDocs parsedDoc, schema *schema.Resource, info *tfbridge.ResourceInfo,
 	isProvider bool) *resourceType {
 
 	return &resourceType{
 		name:       name,
-		doc:        doc,
+		doc:        parsedDocs.Description,
 		isProvider: isProvider,
 		schema:     schema,
 		info:       info,
 		reqprops:   make(map[string]bool),
-		docURL:     docURL,
+		docURL:     parsedDocs.URL,
+		parsedDocs: parsedDocs,
 	}
 }
 
 // resourceFunc is a generated resource function that is exposed to interact with Pulumi objects.
 type resourceFunc struct {
-	name    string
-	doc     string
-	args    []*variable
-	rets    []*variable
-	reqargs map[string]bool
-	argst   *plainOldType
-	retst   *plainOldType
-	schema  *schema.Resource
-	info    *tfbridge.DataSourceInfo
-	docURL  string
+	name       string
+	doc        string
+	args       []*variable
+	rets       []*variable
+	reqargs    map[string]bool
+	argst      *plainOldType
+	retst      *plainOldType
+	schema     *schema.Resource
+	info       *tfbridge.DataSourceInfo
+	docURL     string
+	parsedDocs parsedDoc
 }
 
 func (rf *resourceFunc) Name() string { return rf.name }
@@ -564,7 +567,7 @@ func (g *generator) gatherResource(rawname string,
 	}
 
 	// Create an empty module and associated resource type.
-	res := newResourceType(name, parsedDocs.Description, parsedDocs.URL, schema, info, isProvider)
+	res := newResourceType(name, parsedDocs, schema, info, isProvider)
 
 	args := tfbridge.CleanTerraformSchema(schema.Schema)
 
@@ -708,12 +711,13 @@ func (g *generator) gatherDataSource(rawname string,
 
 	// Build up the function information.
 	fun := &resourceFunc{
-		name:    name,
-		doc:     parsedDocs.Description,
-		reqargs: make(map[string]bool),
-		schema:  ds,
-		info:    info,
-		docURL:  parsedDocs.URL,
+		name:       name,
+		doc:        parsedDocs.Description,
+		reqargs:    make(map[string]bool),
+		schema:     ds,
+		info:       info,
+		docURL:     parsedDocs.URL,
+		parsedDocs: parsedDocs,
 	}
 
 	// Sort the args and return properties so we are ready to go.